requirements.txt
alembic==1.7.6
bleach==4.1.0
blinker==1.4
click==8.0.3
colorama==0.4.4
dnspython==2.2.0
email-validator==1.1.3
Flask==2.0.2
Flask-Admin==1.6.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-MySQLdb==0.2.0
Flask-Principal==0.4.0
Flask-Security-Too==4.1.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
greenlet==1.1.2
idna==3.3
importlib-metadata==4.10.1
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.0.1
mysqlclient==2.1.0
packaging==21.3
passlib==1.7.4
pyparsing==3.0.7
python-decouple==3.6
six==1.16.0
SQLAlchemy==1.4.31
webencodings==0.5.1
Werkzeug==2.0.3
WTForms==3.0.1
zipp==3.7.0
gunicorn==20.1.0
numpy==1.22.2 
pandas==1.4.1 
python-dateutil==2.8.2
